requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2019.9.11
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
generics==3.3.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi==1.1.0
packaging==20.9
pipenv==2018.11.26
psycopg2-binary==2.8.6
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
whitenoise==5.2.0
zipp==3.4.0
